not fully functional image save

+++ app/src/main/java/pt/isec/agileMath/models/Result.kt
@@ -2,6 +2,6 @@ package pt.isec.agileMath.models
 
 data class Result(
     var player: Player = Player(),
-    var score: Int = 0,
-    var totalTime: Int = 0,
+    var score: Long = 0,
+    var totalTime: Long = 0,
 )
